Renamed groupid and package to match promototron required .health vs .med
