+++ mock-ee/pom.xml
@@ -6,7 +6,7 @@
     <artifactId>health-apis-parent</artifactId>
     <version>2.0.5</version>
   </parent>
-  <groupId>gov.va.api.med.mockee</groupId>
+  <groupId>gov.va.api.health.mockee</groupId>
   <artifactId>health-apis-mock-ee</artifactId>
   <version>1.0.17-SNAPSHOT</version>
   <name>health-apis-mock-ee</name>
