+++ mock-ee/src/main/java/gov/va/api/health/mockee/WebServiceConfig.java
@@ -1,4 +1,4 @@
-package gov.va.api.med.mockee;
+package gov.va.api.health.mockee;
 
 import java.util.List;
 import java.util.Properties;
